uv-resolver: switch to "aggressive" forking
This commit switches our forking strategy to one that always forks when
a marker expression is encountered, even if there is only one dependency
specification.

This results in a perf regression but it is believed, at present, to be
necessary for correctness and for stability in the lock file output.