Lift requirement that .egg-info filenames must include version
PR #4533 introduced (almost) spec compliant parsing of .egg-info
filenames, but added the overly strict requirement that the distribution
version must be present. This causes various `uv pip` operations to fail
in environments where there are .egg-info files without a version
component, so loosen this check by making the version component optional
and reading the version from the egg metadata when it is not present.