uv: swap "UTC date" for "local date" in --exclude-newer
Previously, `--exclude-newer` would accept either an RFC 3339 timestamp
or a "UTC date." A UTC date is a somewhat weird interaction point,
because it requires the end user to think about what date it is at
offset 0, which may or may not match their actual date. Instead, we just
accept a "local" date which will be interpreted using their system's
current time zone.

An alternative here would be to get rid of the date option and always
require a RFC 3339 timestamp.