uv/tests: update transformers ecosystem snapshot
It is not clear whether this update is correct or not. Moreover, it's
not clear whether the status quo is correct or not. The problem is that
`transformers` is so big that it's very hard to understand what the
right output is without a deeper investigation.

One thing that is interesting is if fork prioritization is removed in
this PR *and* on `main`, then the differences in this ecosystem test go
away.

We've decided for now to move forward with this update even though we're
uncertain because this PR fixes a few outstanding correctness issues.