Use windows registry to discover python
Our current strategy of parsing the output of `py --list-paths` to get the installed python versions on windows is brittle (#6524, missing `py`, etc.) and it's slow (10ms last time i measured).

Instead, we should behave spec-compliant and read the python versions from the registry following PEP 514.

We're using the official rust-for-windows crates for accessing the registry.

Fixes #1521
Fixes #6524